refactor: move preprocessor.py to core/data/
