+++ simba/core/data/preprocessing_simba.py
@@ -2,9 +2,9 @@ import copy
 
 from simba.core.data.loader_saver import LoaderSaver
 from simba.core.data.loaders import LoadData
+from simba.core.data.preprocessor import Preprocessor
 from simba.core.data.spectrum import SpectrumExt
 from simba.logger_setup import logger
-from simba.preprocessor import Preprocessor
 
 
 class PreprocessingSimba:
