+++ simba/core/models/ordinal/load_data_multitasking.py
@@ -3,6 +3,7 @@ import copy
 import numpy as np
 
 from simba.core.chemistry.tanimoto import Tanimoto
+from simba.core.data.preprocessor import Preprocessor
 from simba.core.models.ordinal.ordinal_classification import (
     OrdinalClassification,
 )
@@ -11,7 +12,6 @@ from simba.core.models.transformers.CustomDatasetMultitasking import (
 )
 from simba.logger_setup import logger
 from simba.molecule_pairs_opt import MoleculePairsOpt
-from simba.preprocessor import Preprocessor
 
 
 class LoadDataMultitasking:
